Issue 34: Move AppearanceBundle into WebBundle
