--- src/Khatovar/Bundle/AppearanceBundle/Resources/config/doctrine/Appearance.orm.yml
@@ -1,45 +1,0 @@
-Khatovar\Bundle\AppearanceBundle\Entity\Appearance:
-    type: entity
-    table: khatovar_web_appearance
-    repositoryClass: Khatovar\Bundle\AppearanceBundle\Entity\AppearanceRepository
-    id:
-        id:
-            type: integer
-            id: true
-            generator:
-                strategy: AUTO
-    fields:
-        name:
-            type: string
-            length: 255
-            column: name
-            nullable: false
-            unique: true
-        slug:
-            type: string
-            length: 128
-            column: slug
-            nullable: false
-            unique: true
-            gedmo:
-                slug:
-                    separator: _
-                    fields:
-                        - name
-        content:
-            type: text
-            column: content
-            nullable: false
-        active:
-            type: boolean
-            column: active
-            nullable: false
-        pageType:
-            type: string
-            length: 255
-            column: page_type
-            nullable: false
-    oneToMany:
-        photos:
-            targetEntity: Khatovar\Bundle\PhotoBundle\Entity\Photo
-            mappedBy: appearance
