--- src/Khatovar/Bundle/AppearanceBundle/Resources/config/form_types.yml
@@ -1,5 +1,0 @@
-services:
-    khatovar_appearance.form.type.appearance:
-        class: Khatovar\Bundle\AppearanceBundle\Form\Type\AppearanceType
-        tags:
-            - { name: form.type }
