--- src/Khatovar/Bundle/AppearanceBundle/Resources/config/routing.yml
@@ -1,48 +1,0 @@
-khatovar_web_appearance:
-    path:     /
-    defaults: { _controller: KhatovarAppearanceBundle:Appearance:index }
-
-khatovar_web_appearance_workshop:
-    path:     /ateliers
-    defaults: { _controller: KhatovarAppearanceBundle:Appearance:workshop }
-
-khatovar_web_appearance_camp:
-    path:     /campement
-    defaults: { _controller: KhatovarAppearanceBundle:Appearance:camp }
-
-khatovar_web_appearance_list:
-    path:     /list
-    defaults: { _controller: KhatovarAppearanceBundle:Appearance:list }
-
-khatovar_web_appearance_show:
-    path:     /prestation/{slug}
-    defaults: { _controller: KhatovarAppearanceBundle:Appearance:show }
-    requirements:
-        slug: '[a-zA-Z0-9-_/]+'
-
-khatovar_web_appearance_new:
-    path:     /new
-    defaults: { _controller: KhatovarAppearanceBundle:Appearance:new }
-
-khatovar_web_appearance_create:
-    path:     /create
-    defaults: { _controller: KhatovarAppearanceBundle:Appearance:create }
-    methods:  POST
-
-khatovar_web_appearance_edit_index:
-    path:     /index/edit
-    defaults: { _controller: KhatovarAppearanceBundle:Appearance:editIndex }
-
-khatovar_web_appearance_edit:
-    path:     /{id}/edit
-    defaults: { _controller: KhatovarAppearanceBundle:Appearance:edit }
-
-khatovar_web_appearance_update:
-    path:     /{id}/update
-    defaults: { _controller: KhatovarAppearanceBundle:Appearance:update }
-    methods:  [POST, PUT]
-
-khatovar_web_appearance_delete:
-    path:     /{id}/delete
-    defaults: { _controller: KhatovarAppearanceBundle:Appearance:delete }
-    methods:  [POST, DELETE]
