--- src/Khatovar/Bundle/AppearanceBundle/Resources/config/twig.yml
@@ -1,7 +1,0 @@
-services:
-    khatovar_appearance.twig.normalizer.link:
-        class: Khatovar\Bundle\AppearanceBundle\Twig\LinkNormalizer
-        arguments:
-            - '@router'
-        tags:
-            - { name: twig.extension }
