--- src/Khatovar/Bundle/AppearanceBundle/Resources/config/validation.yml
@@ -1,12 +1,0 @@
-Khatovar\Bundle\AppearanceBundle\Entity\Appearance:
-    constraints:
-        - Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity: name
-        - Khatovar\Bundle\AppearanceBundle\Validator\Constraints\UniqueActiveCamp: ~
-        - Khatovar\Bundle\AppearanceBundle\Validator\Constraints\UniqueActiveIntro: ~
-    properties:
-        name:
-            - NotBlank: ~
-            - Length:
-                max: 255
-        content:
-            - NotBlank: ~
