--- src/Khatovar/Bundle/AppearanceBundle/Resources/views/Appearance/new.html.twig
@@ -1,12 +1,0 @@
-{% extends 'KhatovarWebBundle::layout.html.twig' %}
-
-{% form_theme form 'KhatovarWebBundle:form:fields.html.twig' %}
-
-{% block inarticle %}
-    <a href="{{ path('khatovar_web_appearance_list') }}">Retour à la liste des prestation</a>
-    <hr />
-
-    <h1>Création d'une nouvelle prestation</h1>
-
-    {% include 'KhatovarAppearanceBundle:Appearance:_form.html.twig' with {'form': form} %}
-{% endblock %}
