+++ src/Khatovar/Bundle/WebBundle/Resources/config/routing.yml
@@ -1,3 +1,7 @@
 khatovar_web_camp:
     path:     /camp
     defaults: { _controller: KhatovarWebBundle:Web:camp }
+
+khatovar_appearance:
+    resource: '@KhatovarWebBundle/Resources/config/routing/appearance.yml'
+    prefix:   /prestations
