+++ src/Khatovar/Bundle/WebBundle/Resources/config/twig.yml
@@ -17,3 +17,10 @@ services:
             - '@khatovar_photo.manager.photo'
         tags:
             - { name: twig.extension }
+
+    khatovar_appearance.twig.normalizer.link:
+        class: Khatovar\Bundle\WebBundle\Twig\LinkNormalizer
+        arguments:
+            - '@router'
+        tags:
+            - { name: twig.extension }
