+++ src/Khatovar/Bundle/WebBundle/Resources/config/validators.yml
@@ -1,0 +1,15 @@
+services:
+    khatovar_appearance.validator.constraint.unique_active_camp:
+        class: Khatovar\Bundle\WebBundle\Validator\Constraints\UniqueActiveCampValidator
+        arguments:
+            - '@doctrine.orm.entity_manager'
+        tags:
+            - { name: validator.constraint_validator, alias: khatovar_unique_active_camp_validator }
+
+    khatovar_appearance.validator.constraint.unique_active_intro:
+        class: Khatovar\Bundle\WebBundle\Validator\Constraints\UniqueActiveIntroValidator
+        arguments:
+            - '@doctrine.orm.entity_manager'
+        tags:
+            - { name: validator.constraint_validator, alias: khatovar_unique_active_intro_validator }
+
