+++ src/Khatovar/Bundle/WebBundle/Resources/views/Appearance/new.html.twig
@@ -1,0 +1,12 @@
+{% extends 'KhatovarWebBundle::layout.html.twig' %}
+
+{% form_theme form 'KhatovarWebBundle:form:fields.html.twig' %}
+
+{% block inarticle %}
+    <a href="{{ path('khatovar_web_appearance_list') }}">Retour à la liste des prestation</a>
+    <hr />
+
+    <h1>Création d'une nouvelle prestation</h1>
+
+    {% include 'KhatovarWebBundle:Appearance:_form.html.twig' with {'form': form} %}
+{% endblock %}
